Add files via upload
This file is the Appendix to be published online with the paper "Methodology for Supervised Optimization of the Construction of Physician Shared-Patient Networks". Section A of the Appendix includes additional dyadic diagnostic measures and conditional dyadic analyses. Section B contains additional empirical results relevant to the application: the demonstration of the minimal risk of overfitting from not forming separate training and test datasets when estimating and comparing the performance of different network construction methods, additional plots of the networks, associations of the Multiplicity factor levels with edge-level diagnostic accuracy, and the fitted Hierarchical P2 model used to evaluate the overall association of the within-hospital shared-patient physician networks with the survey nomination physician networks.